refactor(renderer): split `pipeline_cache` into native/web modules
Moves the cfg-gated `PipelineCache` definitions (`moka` backend on
native, locked map on wasm) into dedicated `pipeline_cache/native.rs`
and `pipeline_cache/web.rs`, mounted via `#[path]`. Each file is
cfg-free.
